PTC-3
Added vSpawn for LHD Vehicles, changed layout of one of the airfields.
Moved the markers of the rSpawn and supply boxes on the airfields.
Removed recover cache mission as its not viable and quite boring imho.
Changed spawnpos of one of the CH-53 to bypass the fireworks that
sometimes happen